Fix 6 defects found in the fourth review round
Both reviewers independently found the first one, which is a regression from
round 3's own fix - the fourth consecutive round in which a fix created the next
defect.

- Round 3 gave the server a tri-state for "has anything ever been collected"
  precisely so a failed read would stop being reported as "the tracker was never
  deployed". The flag was never added to the TypeScript contract, so the client
  still rendered "No SharePoint page views have been collected" and a "Page views
  collected: off" badge on top of a reason saying the opposite - while the
  windowed queries could be succeeding and showing a full report underneath. The
  client now reads it and renders an explicit unknown state.
- The same round fixed the "configuration unreadable" case for the availability
  endpoint only. Every section endpoint still read configuration through a helper
  that returns all-false on failure, so the Overview judgement announced "the
  web-traffic import is switched off" - a deliberate setting - beside real KPIs.
  Readability is now carried through to the judgement.
- The device-mix-over-time chart had no cap on its dimension, unlike the site and
  country stacks. Device names are model-specific for phones (which is exactly
  why the mobile share is deliberately computed from an untruncated query), so on
  a real phone estate this emitted hundreds of bands into a stacked area chart
  whose caption asks whether a responsive-design project is worth funding.
- dbo.cities is keyed on the city NAME alone, so London, UK and London, Ontario
  share a row. The place query grouped on the lookup id and labelled the result
  with MAX(country_id), merging their traffic under whichever country id was
  larger - on a page that shows a country column specifically to tell them apart.
  Cities and regions are now grouped and counted per (place, country).
- Five page exports shared one column set. The slow-page and quiet-page queries
  do not compute entries, exits or bounces - they hard-code zero - so those
  exports shipped placeholder zeros under headings an admin would act on; and the
  entry/exit exports labelled their endpoint counts "Page views". Each export now
  has a column set that matches what its query actually measured.
- The performance judgement told the reader the Technology tab ranks the slowest
  pages. It does not - that table is on Page views.

Also: every TOP (N) now has a deterministic tie-break, so a leaderboard cannot
reshuffle between identical requests; and the client band-boundary test's comment
no longer claims it catches server drift, which it cannot see.

+++ src/AnalyticsEngine/Common/Entities/SpoWebActivity/WebActivityModels.cs
@@ -469,8 +469,12 @@ namespace Common.Entities.SpoWebActivity
         public long Visits { get; set; }
         public int Visitors { get; set; }
         public int Countries { get; set; }
-        public int Cities { get; set; }
-        public int Provinces { get; set; }
+
+        /// <summary>Distinct (city, country) pairs - the city lookup is keyed on name alone.</summary>
+        public long Cities { get; set; }
+
+        /// <summary>Distinct (region, country) pairs.</summary>
+        public long Provinces { get; set; }
 
         /// <summary>Page views whose location the tracker could not resolve.</summary>
         public long UnknownLocationPageViews { get; set; }
